@@ -46,9 +46,9 @@
 		"@popperjs/core": "^2.9.2",
 		"@rollup/plugin-replace": "^5.0.5",
 		"@sveltejs/adapter-netlify": "^3.0.0",
-		"@sveltejs/enhanced-img": "^0.2.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/enhanced-img": "^0.3.9",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@tailwindcss/typography": "^0.5.10",
 		"@types/eslint": "^8.56.7",
 		"@types/js-yaml": "^4.0.9",
@@ -59,7 +59,7 @@
 		"cssnano": "^6.0.1",
 		"eslint": "^9.0.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.36.0",
+		"eslint-plugin-svelte": "^2.45.1",
 		"globals": "^15.0.0",
 		"gray-matter": "^4.0.3",
 		"husky": "^9.0.10",
@@ -68,17 +68,17 @@
 		"postcss": "^8.4.32",
 		"postcss-load-config": "^5.0.2",
 		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
+		"prettier-plugin-svelte": "^3.2.6",
 		"prettier-plugin-tailwindcss": "0.6.8",
 		"random-words": "^2.0.1",
 		"rehype-slug": "^6.0.0",
-		"svelte": "5.0.0-next.154",
-		"svelte-check": "^3.6.0",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
 		"tailwindcss": "^3.3.6",
 		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
+		"typescript": "^5.5.0",
 		"typescript-eslint": "^7.5.0",
-		"vite": "^5.0.3",
+		"vite": "^5.4.4",
 		"vite-plugin-watch-and-run": "^1.7.0",
 		"vitest": "^1.2.0"
 	},
