@@ -13,8 +13,8 @@
 
 	const shows = [Special, Donors1000, Donors1]
 
-	let current = 0
-	let inc = 1
+	let current = $state(0)
+	let inc = $state(1)
 
 	function nextShow() {
 		inc += 1
@@ -30,18 +30,12 @@
 
 <div class="fixed inset-0 dark:bg-black dark:text-white">
 	{#if inc % 2 === 0}
-		<div transition:fade="{{ duration: 1000 }}" class="absolute inset-0">
-			<svelte:component
-				this="{shows[current]}"
-				onEventDone="{() => nextShow()}"
-			/>
+		<div transition:fade={{ duration: 1000 }} class="absolute inset-0">
+			<svelte:component this={shows[current]} onEventDone={() => nextShow()} />
 		</div>
 	{:else}
-		<div transition:fade="{{ duration: 1000 }}" class="absolute inset-0">
-			<svelte:component
-				this="{shows[current]}"
-				onEventDone="{() => nextShow()}"
-			/>
+		<div transition:fade={{ duration: 1000 }} class="absolute inset-0">
+			<svelte:component this={shows[current]} onEventDone={() => nextShow()} />
 		</div>
 	{/if}
 </div>
