@@ -1,5 +1,7 @@
-<script>
+<script lang="ts">
+	type Props = { children?: import("svelte").Snippet }
+	let { children }: Props = $props()
 	import "../app.pcss"
 </script>
 
-<slot></slot>
+{@render children?.()}
